Rbac Feito

+++ backend/controllers/MarcacaoController.php
@@ -6,6 +6,7 @@ use Yii;
 use common\models\MarcacaoConsulta;
 use yii\data\ActiveDataProvider;
 use yii\web\Controller;
+use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 
@@ -70,15 +71,19 @@ class MarcacaoController extends Controller
      */
     public function actionCreate()
     {
-        $model = new MarcacaoConsulta();
+        if (Yii::$app->user->can('criarMarcacaoConsultas')) {
+            $model = new MarcacaoConsulta();
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'idMarcacao_Consulta' => $model->idMarcacao_Consulta, 'Pessoa_idPessoa' => $model->Pessoa_idPessoa]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'idMarcacao_Consulta' => $model->idMarcacao_Consulta, 'Pessoa_idPessoa' => $model->Pessoa_idPessoa]);
+            }
+
+            return $this->render('create', [
+                'model' => $model,
+            ]);
+        }else{
+            throw new ForbiddenHttpException('Não tem permissões');
         }
-
-        return $this->render('create', [
-            'model' => $model,
-        ]);
     }
 
     /**
@@ -91,15 +96,19 @@ class MarcacaoController extends Controller
      */
     public function actionUpdate($idMarcacao_Consulta, $Pessoa_idPessoa)
     {
-        $model = $this->findModel($idMarcacao_Consulta, $Pessoa_idPessoa);
+        if (Yii::$app->user->can('alterarMarcacaoConsultas')) {
+            $model = $this->findModel($idMarcacao_Consulta, $Pessoa_idPessoa);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'idMarcacao_Consulta' => $model->idMarcacao_Consulta, 'Pessoa_idPessoa' => $model->Pessoa_idPessoa]);
+            if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                return $this->redirect(['view', 'idMarcacao_Consulta' => $model->idMarcacao_Consulta, 'Pessoa_idPessoa' => $model->Pessoa_idPessoa]);
+            }
+
+            return $this->render('update', [
+                'model' => $model,
+            ]);
+        }else{
+            throw new ForbiddenHttpException('Não tem permissões');
         }
-
-        return $this->render('update', [
-            'model' => $model,
-        ]);
     }
 
     /**
@@ -112,9 +121,13 @@ class MarcacaoController extends Controller
      */
     public function actionDelete($idMarcacao_Consulta, $Pessoa_idPessoa)
     {
-        $this->findModel($idMarcacao_Consulta, $Pessoa_idPessoa)->delete();
+        if (Yii::$app->user->can('apagarMarcacaoConsultas')) {
+            $this->findModel($idMarcacao_Consulta, $Pessoa_idPessoa)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        }else{
+            throw new ForbiddenHttpException('Não tem permissões');
+        }
     }
 
     /**
